Issue 125 - Migration to AWS Amplify Gen 2
