+++ .gitignore
@@ -44,3 +44,9 @@ amplifytools.xcconfig
 .secret-*
 **.sample
 #amplify-do-not-edit-end
+
+# amplify
+.amplify
+amplify_outputs*
+amplifyconfiguration*
+/amplify/
