+++ package.json
@@ -3,21 +3,21 @@
   "private": true,
   "version": "0.0.0",
   "scripts": {
-    "start": "vite",
+    "dev": "vite",
     "build": "vite build",
     "preview": "vite preview"
   },
   "dependencies": {
     "@aws-amplify/ui-react": "^6.1.12",
-    "aws-amplify": "^6.3.1",
-    "react": "^18.2.0",
+    "aws-amplify": "^6.3.6",
+    "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.2.1"
   },
   "devDependencies": {
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
-    "@vitejs/plugin-react": "^4.3.0",
-    "vite": "^5.2.12"
+    "@vitejs/plugin-react": "^4.3.1",
+    "vite": "^5.2.13"
   }
 }
